feat: Adjustment for role Teknisi

+++ database/migrations/2025_05_24_175327_create_fugitive_emissions_table.php
@@ -14,11 +14,11 @@ return new class extends Migration
         Schema::create('fugitive_emissions', function (Blueprint $table) {
             $table->id();
             $table->string('source_name');
-            $table->float('production_amount');
-            $table->float('emission_factor');
-            $table->float('ch4_emission_ton');
-            $table->float('co2_emission_ton'); // Optional, if needed
-            $table->float('co2e_emission_ton');
+            $table->decimal('production_amount', 15, 10);
+            $table->decimal('emission_factor', 15, 10);
+            $table->decimal('ch4_emission_ton', 15, 10);
+            $table->decimal('co2_emission_ton', 15, 10); // Optional, if needed
+            $table->decimal('co2e_emission_ton', 15, 10);
             $table->date('period');
             $table->unsignedBigInteger('company_id')->default(1); // Default to 1 if no company
             $table->foreign('company_id')->references('id')->on('perusahaans')->onDelete('cascade');
